requirements.txt
snowflake-connector-python==3.7.0